docs(aepp): extract ConnectFrame and ConnectLedger
